[#109] refactor(frontend): extract iks tab helpers and resolve dependency hook warnings
